pyproject.toml
[project]
name = "video-processor-udf"
version = "0.2.0"
description = "Video analytics UDF with AI-powered frame analysis"
requires-python = ">=3.11,<3.12"
dependencies = [
    "couchbase>=4.0.0",
    "ultralytics>=8.0.0",
    "opencv-python-headless>=4.8.0",
    "transformers>=4.40.0",
    "qwen-vl-utils>=0.1.0",
    "sentence-transformers>=2.3.0",
    "accelerate>=0.30.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "autoawq>=0.1.8",
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.0.0",
    "boto3>=1.34.0",
    "Pillow>=10.0.0",
    "numpy<2.0",
]

[project.optional-dependencies]
audio = [
    "moviepy>=1.0.3",
    "openai-whisper>=20231117",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = ["main_udf", "video_ai", "kv_writer"]

